Bug Fixed

+++ src/manager.c
@@ -72,16 +72,15 @@
 #define BUF_SIZE 65535
 #endif
 
-int verbose          = 0;
-char *executable     = "ss-server";
-char *working_dir    = NULL;
+int verbose      = 0;
+char *executable = "ss-server";
+char *working_dir = NULL;
 int working_dir_size = 0;
 
 static struct cork_hash_table *server_table;
 
 #ifndef __MINGW32__
-static int
-setnonblocking(int fd)
+static int setnonblocking(int fd)
 {
     int flags;
     if (-1 == (flags = fcntl(fd, F_GETFL, 0))) {
@@ -92,10 +91,9 @@ setnonblocking(int fd)
 
 #endif
 
-static void
-build_config(char *prefix, struct server *server)
+static void build_config(char *prefix, struct server *server)
 {
-    char *path    = NULL;
+    char *path = NULL;
     int path_size = strlen(prefix) + strlen(server->port) + 20;
 
     path = malloc(path_size);
@@ -105,7 +103,7 @@ build_config(char *prefix, struct server *server)
         if (verbose) {
             LOGE("unable to open config file");
         }
-        ss_free(path);
+        free(path);
         return;
     }
     fprintf(f, "{\n");
@@ -113,11 +111,10 @@ build_config(char *prefix, struct server *server)
     fprintf(f, "\"password\":\"%s\",\n", server->password);
     fprintf(f, "}\n");
     fclose(f);
-    ss_free(path);
+    free(path);
 }
 
-static char *
-construct_command_line(struct manager_ctx *manager, struct server *server)
+static char *construct_command_line(struct manager_ctx *manager, struct server *server)
 {
     static char cmd[BUF_SIZE];
     int i;
@@ -138,12 +135,6 @@ construct_command_line(struct manager_ctx *manager, struct server *server)
         int len = strlen(cmd);
         snprintf(cmd + len, BUF_SIZE - len, " -t %s", manager->timeout);
     }
-#ifdef HAVE_SETRLIMIT
-    if (manager->nofile) {
-        int len = strlen(cmd);
-        snprintf(cmd + len, BUF_SIZE - len, " -n %d", manager->nofile);
-    }
-#endif
     if (manager->user != NULL) {
         int len = strlen(cmd);
         snprintf(cmd + len, BUF_SIZE - len, " -a %s", manager->user);
@@ -168,10 +159,6 @@ construct_command_line(struct manager_ctx *manager, struct server *server)
         int len = strlen(cmd);
         snprintf(cmd + len, BUF_SIZE - len, " --fast-open");
     }
-    if (manager->mtu) {
-        int len = strlen(cmd);
-        snprintf(cmd + len, BUF_SIZE - len, " --mtu %d", manager->mtu);
-    }
     for (i = 0; i < manager->nameserver_num; i++) {
         int len = strlen(cmd);
         snprintf(cmd + len, BUF_SIZE - len, " -d %s", manager->nameservers[i]);
@@ -188,8 +175,7 @@ construct_command_line(struct manager_ctx *manager, struct server *server)
     return cmd;
 }
 
-static char *
-get_data(char *buf, int len)
+static char *get_data(char *buf, int len)
 {
     char *data;
     int pos = 0;
@@ -204,8 +190,7 @@ get_data(char *buf, int len)
     return data;
 }
 
-static char *
-get_action(char *buf, int len)
+static char *get_action(char *buf, int len)
 {
     char *action;
     int pos = 0;
@@ -224,8 +209,7 @@ get_action(char *buf, int len)
     return action;
 }
 
-static struct server *
-get_server(char *buf, int len)
+static struct server *get_server(char *buf, int len)
 {
     char *data = get_data(buf, len);
     char error_buf[512];
@@ -271,8 +255,7 @@ get_server(char *buf, int len)
     return server;
 }
 
-static int
-parse_traffic(char *buf, int len, char *port, uint64_t *traffic)
+static int parse_traffic(char *buf, int len, char *port, uint64_t *traffic)
 {
     char *data = get_data(buf, len);
     char error_buf[512];
@@ -305,8 +288,7 @@ parse_traffic(char *buf, int len, char *port, uint64_t *traffic)
     return 0;
 }
 
-static void
-add_server(struct manager_ctx *manager, struct server *server)
+static void add_server(struct manager_ctx *manager, struct server *server)
 {
     bool new = false;
     cork_hash_table_put(server_table, (void *)server->port, (void *)server, &new, NULL, NULL);
@@ -317,8 +299,7 @@ add_server(struct manager_ctx *manager, struct server *server)
     }
 }
 
-static void
-kill_server(char *prefix, char *pid_file)
+static void kill_server(char *prefix, char *pid_file)
 {
     char *path = NULL;
     int pid, path_size = strlen(prefix) + strlen(pid_file) + 2;
@@ -329,7 +310,7 @@ kill_server(char *prefix, char *pid_file)
         if (verbose) {
             LOGE("unable to open pid file");
         }
-        ss_free(path);
+        free(path);
         return;
     }
     if (fscanf(f, "%d", &pid) != EOF) {
@@ -337,11 +318,10 @@ kill_server(char *prefix, char *pid_file)
     }
     fclose(f);
     remove(path);
-    ss_free(path);
+    free(path);
 }
 
-static void
-stop_server(char *prefix, char *port)
+static void stop_server(char *prefix, char *port)
 {
     char *path = NULL;
     int pid, path_size = strlen(prefix) + strlen(port) + 20;
@@ -352,18 +332,17 @@ stop_server(char *prefix, char *port)
         if (verbose) {
             LOGE("unable to open pid file");
         }
-        ss_free(path);
+        free(path);
         return;
     }
     if (fscanf(f, "%d", &pid) != EOF) {
         kill(pid, SIGTERM);
     }
     fclose(f);
-    ss_free(path);
+    free(path);
 }
 
-static void
-remove_server(char *prefix, char *port)
+static void remove_server(char *prefix, char *port)
 {
     char *old_port            = NULL;
     struct server *old_server = NULL;
@@ -377,8 +356,7 @@ remove_server(char *prefix, char *port)
     stop_server(prefix, port);
 }
 
-static void
-update_stat(char *port, uint64_t traffic)
+static void update_stat(char *port, uint64_t traffic)
 {
     void *ret = cork_hash_table_get(server_table, (void *)port);
     if (ret != NULL) {
@@ -387,8 +365,7 @@ update_stat(char *port, uint64_t traffic)
     }
 }
 
-static void
-manager_recv_cb(EV_P_ ev_io *w, int revents)
+static void manager_recv_cb(EV_P_ ev_io *w, int revents)
 {
     struct manager_ctx *manager = (struct manager_ctx *)w;
     socklen_t len;
@@ -510,8 +487,7 @@ ERROR_MSG:
     }
 }
 
-static void
-signal_cb(EV_P_ ev_signal *w, int revents)
+static void signal_cb(EV_P_ ev_signal *w, int revents)
 {
     if (revents & EV_SIGNAL) {
         switch (w->signum) {
@@ -522,8 +498,7 @@ signal_cb(EV_P_ ev_signal *w, int revents)
     }
 }
 
-int
-create_server_socket(const char *host, const char *port)
+int create_server_socket(const char *host, const char *port)
 {
     struct addrinfo hints;
     struct addrinfo *result, *rp, *ipv4v6bindall;
@@ -597,8 +572,7 @@ create_server_socket(const char *host, const char *port)
     return server_sock;
 }
 
-int
-main(int argc, char **argv)
+int main(int argc, char **argv)
 {
     int i, c;
     int pid_flags         = 0;
@@ -615,11 +589,6 @@ main(int argc, char **argv)
     int auth      = 0;
     int fast_open = 0;
     int mode      = TCP_ONLY;
-    int mtu       = 0;
-
-#ifdef HAVE_SETRLIMIT
-static int nofile = 0;
-#endif
 
     int server_num = 0;
     char *server_host[MAX_REMOTE_NUM];
@@ -631,12 +600,11 @@ static int nofile = 0;
 
     int option_index                    = 0;
     static struct option long_options[] = {
-        { "fast-open",       no_argument,       0, 0 },
-        { "acl",             required_argument, 0, 0 },
+        { "fast-open"      , no_argument      , 0, 0 },
+        { "acl"            , required_argument, 0, 0 },
         { "manager-address", required_argument, 0, 0 },
-        { "executable",      required_argument, 0, 0 },
-        { "mtu",             required_argument, 0, 0 },
-        { "help",            no_argument,       0, 0 },
+        { "executable"     , required_argument, 0, 0 },
+        { "help"           , no_argument      , 0, 0 },
         {                 0,                 0, 0, 0 }
     };
 
@@ -644,7 +612,7 @@ static int nofile = 0;
 
     USE_TTY();
 
-    while ((c = getopt_long(argc, argv, "f:s:l:k:t:m:c:i:d:a:n:huUvA",
+    while ((c = getopt_long(argc, argv, "f:s:l:k:t:m:c:i:d:a:huUvA",
                             long_options, &option_index)) != -1)
         switch (c) {
         case 0:
@@ -657,9 +625,6 @@ static int nofile = 0;
             } else if (option_index == 3) {
                 executable = optarg;
             } else if (option_index == 4) {
-                mtu = atoi(optarg);
-                LOGI("set MTU to %d", mtu);
-            } else if (option_index == 5) {
                 usage();
                 exit(EXIT_SUCCESS);
             }
@@ -711,14 +676,8 @@ static int nofile = 0;
         case 'A':
             auth = 1;
             break;
-#ifdef HAVE_SETRLIMIT
-        case 'n':
-            nofile = atoi(optarg);
-            break;
-#endif
         case '?':
             // The option character is not recognized.
-            LOGE("Unrecognized option: %s", optarg);
             opterr = 1;
             break;
         }
@@ -758,14 +717,6 @@ static int nofile = 0;
         if (mode == TCP_ONLY) {
             mode = conf->mode;
         }
-        if (mtu == 0) {
-            mtu = conf->mtu;
-        }
-#ifdef HAVE_SETRLIMIT
-        if (nofile == 0) {
-            nofile = conf->nofile;
-        }
-#endif
     }
 
     if (server_num == 0) {
@@ -836,10 +787,6 @@ static int nofile = 0;
     manager.host_num        = server_num;
     manager.nameservers     = nameservers;
     manager.nameserver_num  = nameserver_num;
-    manager.mtu             = mtu;
-#ifdef HAVE_SETRLIMIT
-    manager.nofile          = nofile;
-#endif
 
     // initialize ev loop
     struct ev_loop *loop = EV_DEFAULT;
@@ -851,14 +798,14 @@ static int nofile = 0;
 
     struct passwd *pw   = getpwuid(getuid());
     const char *homedir = pw->pw_dir;
-    working_dir_size = strlen(homedir) + 15;
-    working_dir      = malloc(working_dir_size);
+    working_dir_size = strlen(homedir)+15;
+    working_dir = malloc(working_dir_size);
     snprintf(working_dir, working_dir_size, "%s/.shadowsocks", homedir);
 
     int err = mkdir(working_dir, S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
     if (err != 0 && errno != EEXIST) {
         ERROR("mkdir");
-        ss_free(working_dir);
+        free(working_dir);
         FATAL("unable to create working directory");
     }
 
@@ -871,13 +818,12 @@ static int nofile = 0;
             size_t len = strlen(ep->d_name);
             if (strcmp(ep->d_name + len - 3, "pid") == 0) {
                 kill_server(working_dir, ep->d_name);
-                if (verbose)
-                    LOGI("kill %s", ep->d_name);
+                if (verbose) LOGI("kill %s", ep->d_name);
             }
         }
-        closedir(dp);
+        closedir (dp);
     } else {
-        ss_free(working_dir);
+        free(working_dir);
         FATAL("Couldn't open the directory");
     }
 
@@ -900,7 +846,7 @@ static int nofile = 0;
         struct sockaddr_un svaddr;
         sfd = socket(AF_UNIX, SOCK_DGRAM, 0);       /*  Create server socket */
         if (sfd == -1) {
-            ss_free(working_dir);
+            free(working_dir);
             FATAL("socket");
         }
 
@@ -908,7 +854,7 @@ static int nofile = 0;
 
         if (remove(manager_address) == -1 && errno != ENOENT) {
             ERROR("bind");
-            ss_free(working_dir);
+            free(working_dir);
             exit(EXIT_FAILURE);
         }
 
@@ -918,13 +864,13 @@ static int nofile = 0;
 
         if (bind(sfd, (struct sockaddr *)&svaddr, sizeof(struct sockaddr_un)) == -1) {
             ERROR("bind");
-            ss_free(working_dir);
+            free(working_dir);
             exit(EXIT_FAILURE);
         }
     } else {
         sfd = create_server_socket(ip_addr.host, ip_addr.port);
         if (sfd == -1) {
-            ss_free(working_dir);
+            free(working_dir);
             FATAL("socket");
         }
     }
@@ -957,7 +903,7 @@ static int nofile = 0;
 
     ev_signal_stop(EV_DEFAULT, &sigint_watcher);
     ev_signal_stop(EV_DEFAULT, &sigterm_watcher);
-    ss_free(working_dir);
+    free(working_dir);
 
     return 0;
 }
